Refactoring
Added a new class called Alarm. This class has the Buzzer and the two Leds.
Now the code on the main program is simpler

+++ Buzzer.cpp
@@ -7,7 +7,7 @@ class Buzzer {
     ezBuzzer buzzer;
 
   public:
-  Buzzer::Buzzer(int pin):buzzer(pin) {
+  Buzzer::Buzzer(byte pin):buzzer(pin) {
     this->pin = pin;
   }
